Sepparate resolvers in separate files and in src folder
